requirements.txt
langchain
python-dotenv
openai
PyPDF2
faiss
langchain_openai
faiss-cpu
